Chromium Mac and Linux symbols are not archived atm, but I'll run the Chromebot for them anyway. If there are any crashes, dumps will be provided but no crash report.
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11953006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@178379 0039d316-1c4b-4281-b951-d872f2087c98

+++ scripts/slave/get_build_for_chromebot.py
@@ -30,6 +30,8 @@ import urllib
 import urllib2
 import zipfile
 
+from slave import slave_utils
+
 
 def RemovePath(path):
   """Remove the given dir."""
@@ -224,9 +226,10 @@ class GetBuild(object):
 
   def DownloadAndExtractFiles(self):
     """Download and extract files."""
+    exit_code = 0
     if not DoesURLExist(self._chrome_zip_url):
       print 'URL does not exist : ' + self._chrome_zip_url
-      return False
+      return slave_utils.ERROR_EXIT_CODE
 
     os.makedirs(self._target_dir)
     os.chmod(self._target_dir, 0755)
@@ -243,12 +246,16 @@ class GetBuild(object):
     file_path = os.path.join(self._target_dir, self._test_name)
     urllib.urlretrieve(self._test_url, file_path)
 
-    # Download and extract breakpad symbols.
-    print 'Downloading URL: ' + self._symbol_url
-    file_path = os.path.join(self._target_dir, self._symbol_name)
-    urllib.urlretrieve(self._symbol_url, file_path)
-    if self._options.extract:
-      Extract(file_path, self._symbol_dir)
+    # Download and extract breakpad symbols. Skip if they don't exist.
+    if DoesURLExist(self._symbol_url):
+      print 'Downloading URL: ' + self._symbol_url
+      file_path = os.path.join(self._target_dir, self._symbol_name)
+      urllib.urlretrieve(self._symbol_url, file_path)
+      if self._options.extract:
+        Extract(file_path, self._symbol_dir)
+    else:
+      print '\nWARNING: No breakpad symbol available.'
+      exit_code = slave_utils.WARNING_EXIT_CODE
 
     # Download ffmpegsumo.so.
     if self._ffmpegsumo_name:
@@ -259,7 +266,8 @@ class GetBuild(object):
     for path, _, fnames in os.walk(self._target_dir):
       for fname in fnames:
         os.chmod(os.path.join(path, fname), 0755)
-    return True
+
+    return exit_code
 
   def Main(self):
     """main() routine for GetBuild.  Fetch everything.
@@ -267,14 +275,13 @@ class GetBuild(object):
     Returns:
       Value suitable for process exit code (e.g. 0 on success).
     """
-    if (not self.ProcessArgs() or
-        not self.CleanUp() or
-        not self.DownloadAndExtractFiles()):
-      return 1
+    if not self.ProcessArgs() or not self.CleanUp():
+      return slave_utils.ERROR_EXIT_CODE
 
     # See scripts/master/factory/commands.py's SetBuildPropertyShellCommand
     print 'BUILD_PROPERTY build_id=%s' % self._build_id
-    return 0
+
+    return self.DownloadAndExtractFiles()
 
 
 def main():
